+++ Library/src/main/java/com/training/library/services/BookStatusService.java
@@ -10,12 +10,14 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.training.library.dto.request.BookStatusRequestDto;
 import com.training.library.dto.request.FilterDto;
 import com.training.library.dto.response.BookStatusResponseDto;
 import com.training.library.dto.response.CustomBaseResponseDto;
 import com.training.library.entity.BookStatus;
+import com.training.library.entity.Location;
 import com.training.library.repositories.BookStatusRepository;
 
 @Service
@@ -28,6 +30,8 @@ public class BookStatusService {
 	private Environment env;
 	@Autowired
 	private LocationService locationService;
+	@Autowired
+	private BookDetailsService bookDetailsService;
 
 	public BookStatusResponseDto findBookStatus(Long id) {
 		Optional<BookStatusResponseDto> bookStatus = bookStatusRepository.findByBookStatusIdAndDeletedAtIsNull(id);
@@ -37,8 +41,33 @@ public class BookStatusService {
 		return null;
 	}
 
+	@Transactional
 	public ResponseEntity<CustomBaseResponseDto> deleteBookStatus(Long id) {
-		bookStatusRepository.deleteByBookStatusId(id);
+		BookStatus bookStatus = findBookById(id);
+		if (Boolean.TRUE.equals(bookStatus.isAvailable())) {
+			Location locationOld = locationService.findLocationById(bookStatus.getLocation().getLocationId());
+			locationOld.setIsAvailable(true);
+			locationService.updateLocationAvailability(locationOld);
+			bookDetailsService.setAvailableCopies(bookStatus.getBookDetails(), "remove");
+			bookStatusRepository.deleteByBookStatusId(id);
+		} else {
+			throw new RuntimeException("For delete book, book really required at location");
+		}
+		return ResponseEntity.ok(new CustomBaseResponseDto(env.getRequiredProperty(OPERATION_SUCCESS)));
+
+	}
+	
+	@Transactional
+	public ResponseEntity<CustomBaseResponseDto> deleteBookStatus(BookStatus bookStatus) {
+		if (Boolean.TRUE.equals(bookStatus.isAvailable())) {
+			Location locationOld = locationService.findLocationById(bookStatus.getLocation().getLocationId());
+			locationOld.setIsAvailable(true);
+			locationService.updateLocationAvailability(locationOld);
+			bookDetailsService.setAvailableCopies(bookStatus.getBookDetails(), "remove");
+			bookStatusRepository.deleteByBookStatusId(bookStatus.getBookStatusId());
+		} else {
+			throw new RuntimeException("For delete book, book really required at location");
+		}
 		return ResponseEntity.ok(new CustomBaseResponseDto(env.getRequiredProperty(OPERATION_SUCCESS)));
 
 	}
@@ -78,6 +107,54 @@ public class BookStatusService {
 
 	public void updateBookStatusAvailability(BookStatus bookStatus) {
 		bookStatusRepository.save(bookStatus);
+	}
+
+	public BookStatus findByLocationId(Long locationId) {
+		Optional<BookStatus> bookStatusOptional = bookStatusRepository.findByLocation_LocationId(locationId);
+		if (bookStatusOptional.isPresent()) {
+			return bookStatusOptional.get();
+		}
+		return null;
+	}
+
+	public ResponseEntity<CustomBaseResponseDto> getCountByLocation(Long id) {
+		Long counter = bookStatusRepository.countByLocation_LocationIdAndDeletedAtIsNull(id);
+		return ResponseEntity.ok(new CustomBaseResponseDto(String.valueOf(counter)));
+	}
+
+	public ResponseEntity<CustomBaseResponseDto> getCountByShelf(Long id) {
+		Long counter = bookStatusRepository.countByLocation_Shelf_ShelfIdAndDeletedAtIsNull(id);
+		return ResponseEntity.ok(new CustomBaseResponseDto(String.valueOf(counter)));
+	}
+
+	public ResponseEntity<CustomBaseResponseDto> getCountBySection(Long id) {
+		Long counter = bookStatusRepository.countByLocation_Shelf_Section_SectionIdAndDeletedAtIsNull(id);
+		return ResponseEntity.ok(new CustomBaseResponseDto(String.valueOf(counter)));
+	}
+
+	public ResponseEntity<CustomBaseResponseDto> getCountByFloor(Long id) {
+		Long counter = bookStatusRepository.countByLocation_Shelf_Section_Floor_FloorIdAndDeletedAtIsNull(id);
+		return ResponseEntity.ok(new CustomBaseResponseDto(String.valueOf(counter)));
+	}
+
+	public Page<BookStatusResponseDto> findAllBooksByLocation(Long id, FilterDto dto) {
+		Pageable pageble = PageRequest.of(dto.getPageNumber(), dto.getPageSize());
+		return bookStatusRepository.findAllByLocationId(id, pageble);
+	}
+
+	public Page<BookStatusResponseDto> findAllBooksByShelf(Long id, FilterDto dto) {
+		Pageable pageble = PageRequest.of(dto.getPageNumber(), dto.getPageSize());
+		return bookStatusRepository.findAllByShelfId(id, pageble);
+	}
+
+	public Page<BookStatusResponseDto> findAllBooksByFloor(Long id, FilterDto dto) {
+		Pageable pageble = PageRequest.of(dto.getPageNumber(), dto.getPageSize());
+		return bookStatusRepository.findAllByFloorId(id, pageble);
+	}
+
+	public Page<BookStatusResponseDto> findAllBooksBySection(Long id, FilterDto dto) {
+		Pageable pageble = PageRequest.of(dto.getPageNumber(), dto.getPageSize());
+		return bookStatusRepository.findAllBySectionId(id, pageble);
 	}
 
 }
